lab-work/lab work 4/requirements.txt
loguru==0.6.0
python-dotenv==0.21.0
bcrypt==4.0.1

